Add basic elixir testing workflow
Creates a workflow for testing each new or modified exercise individually,
checking for modification against its own branch head.

Currently, it uses Elixir 1.14.x and OTP 25.x for setting erlef/setup-beam up.

Each selected exercise is tested in a separate job, which might cause problem
if there are more than 256
(https://docs.github.com/en/actions/using-jobs/using-a-matrix-for-your-jobs#using-a-matrix-strategy).
This is unlikely as long as each push modifies only a couple of solutions.